api/change: keep track of workspace_id
--- sturdy ---
user_id: cc33c18b-27e5-470d-8589-f9391ca37590
view_id: 2c064d23-2410-4e86-939f-91c56dae2fae
workspace_id: 6ff84ac8-3826-42bc-840d-6a633319a3e8
