+++ api/pkg/change/db/db.go
@@ -4,20 +4,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/lib/pq"
-
 	"getsturdy.com/api/pkg/change"
 
 	"github.com/jmoiron/sqlx"
+	"github.com/lib/pq"
 )
-
-type Repository interface {
-	Get(ctx context.Context, id change.ID) (*change.Change, error)
-	ListByIDs(ctx context.Context, ids ...change.ID) ([]*change.Change, error)
-	GetByCommitID(ctx context.Context, commitID, codebaseID string) (*change.Change, error)
-	Insert(ctx context.Context, ch change.Change) error
-	Update(ctx context.Context, ch change.Change) error
-}
 
 func NewRepo(db *sqlx.DB) Repository {
 	return &repo{db: db}
@@ -29,7 +20,7 @@ type repo struct {
 
 func (r *repo) Get(ctx context.Context, id change.ID) (*change.Change, error) {
 	var res change.Change
-	err := r.db.GetContext(ctx, &res, `SELECT id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id
+	err := r.db.GetContext(ctx, &res, `SELECT id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id, workspace_id
 		FROM changes
 		WHERE id = $1`, id)
 	if err != nil {
@@ -40,7 +31,7 @@ func (r *repo) Get(ctx context.Context, id change.ID) (*change.Change, error) {
 
 func (r *repo) GetByCommitID(ctx context.Context, commitID, codebaseID string) (*change.Change, error) {
 	var res change.Change
-	err := r.db.GetContext(ctx, &res, `SELECT id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id
+	err := r.db.GetContext(ctx, &res, `SELECT id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id, workspace_id
 		FROM changes
 		WHERE commit_id = $1 AND codebase_id = $2`, commitID, codebaseID)
 	if err != nil {
@@ -51,8 +42,8 @@ func (r *repo) GetByCommitID(ctx context.Context, commitID, codebaseID string) (
 
 func (r *repo) Insert(ctx context.Context, ch change.Change) error {
 	_, err := r.db.NamedExecContext(ctx, `INSERT INTO changes
-		(id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id)
-		VALUES(:id, :codebase_id, :title, :updated_description, :user_id, :git_creator_name, :git_creator_email, :created_at, :git_created_at, :commit_id, :parent_change_id)
+		(id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id, workspace_id)
+		VALUES(:id, :codebase_id, :title, :updated_description, :user_id, :git_creator_name, :git_creator_email, :created_at, :git_created_at, :commit_id, :parent_change_id, :workspace_id)
     	`, &ch)
 	if err != nil {
 		return fmt.Errorf("failed to insert: %w", err)
@@ -70,7 +61,8 @@ func (r *repo) Update(ctx context.Context, ch change.Change) error {
     	    created_at = :created_at,
     	    git_created_at = :git_created_at,
 			commit_id = :commit_id,
-    	    parent_change_id = :parent_change_id
+    	    parent_change_id = :parent_change_id,
+			workspace_id = :workspace_id
     	WHERE id = :id`, &ch)
 	if err != nil {
 		return fmt.Errorf("failed to update change: %w", err)
@@ -82,7 +74,7 @@ func (r *repo) ListByIDs(ctx context.Context, ids ...change.ID) ([]*change.Chang
 	var res []*change.Change
 	err := r.db.SelectContext(ctx, &res, `
 		SELECT
-			id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id
+			id, codebase_id, title, updated_description, user_id, git_creator_name, git_creator_email, created_at, git_created_at, commit_id, parent_change_id, workspace_id
 		FROM
 			changes
 		WHERE
